--- Godeps/Godeps.json
@@ -1,26 +1,0 @@
-{
-	"ImportPath": "github.com/cloudfoundry-incubator/spiff",
-	"GoVersion": "go1.4.1",
-	"Deps": [
-		{
-			"ImportPath": "github.com/codegangsta/cli",
-			"Comment": "1.2.0-95-g9b2bd2b",
-			"Rev": "9b2bd2b3489748d4d0a204fa4eb2ee9e89e0ebc6"
-		},
-		{
-			"ImportPath": "github.com/onsi/ginkgo",
-			"Comment": "v1.1.0-42-gdbb5c6c",
-			"Rev": "dbb5c6caf33238b57facc1d975b1aaca6b90288c"
-		},
-		{
-			"ImportPath": "github.com/onsi/gomega",
-			"Comment": "v1.0-39-gc853a76",
-			"Rev": "c853a76367c700fe446c27bc7e6fc8209ee4e0a1"
-		},
-		{
-			"ImportPath": "github.com/pointlander/peg",
-			"Comment": "non released snapshot",
-			"Rev": "1d0268dfff9bca9748dc9105a214ace2f5c594a8"
-		}
-	]
-}
